requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
defusedxml==0.7.1
Django==4.2.6
django-admin-honeypot==1.1.0
django-cors-headers==4.3.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
djangorestframework-yaml==2.0.0
drf-yasg==1.21.7
Faker==19.11.0
inflection==0.5.1
Markdown==3.5
packaging==23.2
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
validate-docbr==1.10.0
